+++ src/backend/joanie/tests/core/test_models_organization.py
@@ -61,6 +61,7 @@ class OrganizationModelsTestCase(BaseAPITestCase):
                 "download_quote": False,
                 "confirm_quote": False,
                 "confirm_bank_transfer": False,
+                "can_submit_for_signature_batch_order": False,
             },
         )
 
@@ -80,6 +81,7 @@ class OrganizationModelsTestCase(BaseAPITestCase):
                 "download_quote": False,
                 "confirm_quote": False,
                 "confirm_bank_transfer": False,
+                "can_submit_for_signature_batch_order": False,
             },
         )
 
@@ -99,6 +101,7 @@ class OrganizationModelsTestCase(BaseAPITestCase):
                 "download_quote": True,
                 "confirm_quote": True,
                 "confirm_bank_transfer": True,
+                "can_submit_for_signature_batch_order": True,
             },
         )
 
@@ -118,6 +121,7 @@ class OrganizationModelsTestCase(BaseAPITestCase):
                 "download_quote": False,
                 "confirm_quote": False,
                 "confirm_bank_transfer": False,
+                "can_submit_for_signature_batch_order": False,
             },
         )
 
@@ -140,6 +144,7 @@ class OrganizationModelsTestCase(BaseAPITestCase):
                 "download_quote": False,
                 "confirm_quote": False,
                 "confirm_bank_transfer": False,
+                "can_submit_for_signature_batch_order": False,
             },
         )
 
@@ -163,6 +168,7 @@ class OrganizationModelsTestCase(BaseAPITestCase):
                 "download_quote": False,
                 "confirm_quote": False,
                 "confirm_bank_transfer": False,
+                "can_submit_for_signature_batch_order": False,
             },
         )